Bug fixed.
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@4969 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/grep_engine.h
@@ -30,7 +30,7 @@ public:
     
 private:
    
-    bool finalLineIsUnterminated(char * fileBuffer, int fileSize) const;
+    bool finalLineIsUnterminated(char * fileBuffer, size_t fileSize) const;
 
     GrepFunctionType mMainFcn;
